requirements.txt
Flask
PyPDF2
Pillow
tabula-py
pandas
pdf2docx
Werkzeug
